ocr-eval/requirements.txt
nltk==3.9.1
torchmetrics==1.6.1
openai==1.61.1
paddleocr>=2.0.1
python-Levenshtein==0.26.1
sacrebleu==2.5.1
datasets==3.2.0
google-genai>=1.0.0
surya-ocr==0.11.0
qwen-vl-utils==0.0.10
accelerate==1.4.0
google-genai==1.0.0
torch==2.6.0
torchvision==0.21.0
transformers==4.49.0
regex==2024.11.6